Crate sewup

Sewup - Second state eWasm Utility Program A library to help you sew up your Ethereum project with Rust and just like develop in a common backend.

Use the crate with different feature to use the high level api just enable the features you want to use.

KV feature helps you develop contract as key value database

sewup = { version = "*", features = ['kv'] }
sewup-derive = { version = "*", features = ['kv']  }

RDB feature helps you develop contract as rdb database

sewup = { version = "*", features = ['rdb'] }
sewup-derive = { version = "*", features = ['rdb']  }

Re-exports

pub use token::*;

pub use kv::*;

pub use rdb::*;

pub use bincode;

Modules

errors

kv

help you storage thing as key value object pair kv feature provides a simple way to key/value store things into ethereum runtime. It aims to be lightweight and with a nice high level interface. Store is an abstract storage instance from one account in one block There can be more than one bucket in a store. Besides, store can improt data from the specific block. Different bucket can defined different kind of key value storage pair.

primitives

rdb

help you storage thing as records in tables rdb feature provides a simple way to store things with relations into ethereum runtime.

runtimes

The run time helps user to setup the contract testing environment

token

help you build up you contract to handle tokens (experimental)

types

The basic types for storage in low level, and also easiler to used for bytes and string. Sewup using Raw and Row as basic unit to storage the data

utils

Macros

ewasm_dbg

Enums

Value

Traits

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

DeserializeTrait

A data structure that can be deserialized from any data format supported by Serde.

SerializeTrait

A data structure that can be serialized into any data format supported by Serde.

Functions

to_value

Type Definitions

Result

Result<T, Error>

Derive Macros

Deserialize

DeserializeTrait

Serialize

SerializeTrait