Module sewup::rdb

help you storage thing as records in tables rdb feature provides a simple way to store things with relations into ethereum runtime.

Modules

errors

traits
traits for Record, which the item in an table

Structs

Db
DB is a storage space for an account in a specific block. We can import the storage from a past block, and we only commit the storage into the latest block.

Table

TableInfo
Metadata of table

Enums

Feature
DB feature flag to enable the different feature for db