Struct sewup::rdb::Db

pub struct Db { /* fields omitted */ }

DB is a storage space for an account in a specific block. We can import the storage from a past block, and we only commit the storage into the latest block.

Storage map

| 0th ~ 31th bytes | dynamic size | dynamic size              | dynamic size |
|------------------|--------------|---------------------------|--------------|
| DB header        | Table info   | Table data of first table | ...          |

DB Header

The fist 32 bytes are reserved as header of the store,

| 0th            | 1st          | 2nd ~ 3rd         | ... | 28th ~ 31st              |
|----------------|--------------|-------------------|-----|--------------------------|
| Sewup Features | version (BE) | RDB Features (LE) | -   | length of TableInfo (BE) |

Base on the features, the storage may have different encoding in to binary

Implementations

impl Db

pub fn new() -> Result<Self>

pub fn version(&self) -> u8

The version of current storage

pub fn features(&self) -> Vec<Feature>

The feature enabled in current storage

pub fn create_table<T: SerializeTrait + Default + Sized + Record>(
&mut self
) -> Result<()>

create table for storage

pub fn table<T: SerializeTrait + Default + Sized + Record>(
self
) -> Result<Table<T>>

get table with date loaded

pub fn drop_table<T>(&mut self)

drop table

pub fn table_length(&self) -> usize

get the numbers of tables

pub fn table_info<T>(&self) -> Option<TableInfo>

pub fn load(_block_height: Option<i64>) -> Result<Self>

pub fn commit(&self) -> Result<()>