Crate sewup

Sewup - Second state eWasm Utility Program A library to help you sew up your Ethereum project with Rust and just like develop in a common backend.

Use the crate with different feature to use the high level api just enable the features you want to use.

Re-exports

pub use kv::*;

pub use bincode;

Modules

errors

kv
kv feature provides a simple way to key/value store things into evm It aims to be lightweight and with a nice high level interface. Store is an abstract storage instance from one account in one block There can be more than one bucket in a store. Besides, store can improt data from the specific block. Different bucket can defined different kind of key value storage pair.

primitives

runtimes

types

utils