Module sewup::runtimes::traits

RT Trait This trait help you abstract any VM into application runtime layer, such that you can easily use your VM with SewUp libs.

Besides, the VMErrors, VMmessage VMresult are rust style wrarp for evm_error, evm_message and evm_result

Structs

VMMessage

VMMessageBuilder

VMResult

Enums

Flags

VmError

Traits

RT